Manejo de States terminado - Quedaria agregar otros States, ej: Menu - Eloy
