--- HelloWorld-Box2d/core/src/com/helloworld/box2dprueba/states/PlayState.java
@@ -1,44 +1,0 @@
-package com.helloworld.box2dprueba.states;
-
-import box2dLight.ConeLight;
-import box2dLight.PointLight;
-import box2dLight.RayHandler;
-import com.badlogic.gdx.ApplicationAdapter;
-import com.badlogic.gdx.ApplicationListener;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.Animation;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
-import com.badlogic.gdx.maps.tiled.TiledMap;
-import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
-import com.badlogic.gdx.physics.box2d.Body;
-import com.badlogic.gdx.physics.box2d.Box2DDebugRenderer;
-import com.badlogic.gdx.physics.box2d.World;
-
-public class PlayState extends State {
-
-
-    public PlayState(GameStateManager gsm) {
-        super(gsm);
-    }
-
-    @Override
-    public void handleInput() {
-
-    }
-
-    @Override
-    public void update(float delta) {
-
-    }
-
-    @Override
-    public void render(SpriteBatch sb) {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
-}
